Add VPN LaunchAgent removal attribute and relatable code
My team uses the `cpe_anyconnect` cookbook/resource, and it does most of what we need it to do. Our organization needs the ability to opt out of starting/creating the LaunchAgent for the GUI. We are only using the Umbrella Security Module of AnyConnect and have another VPN solution in place.

This PR adds a new attribute and some code that allows the ability to skip over starting, or restarting, the LaunchAgent for the GUI.